requirements.txt
langchain
openai
huggingface_hub
langchain_community
langchain-google-genai
python-dotenv
streamlit